+++ behave/configuration.py
@@ -69,6 +69,12 @@ options = [
      dict(action='store_true',
           help="Invokes formatters without executing the steps.")),
 
+    (('-D', '--define'),
+     dict(action='append', dest='userdata',
+          help="""Define custom settings available to steps in
+                  context.config.userdata. Example: --define foo=bar
+                  puts the string "bar" into context.config.userdata.foo.""")),
+
     (('-e', '--exclude'),
      dict(metavar="PATTERN", dest='exclude_re',
           help="""Don't run feature files matching regular expression
@@ -392,6 +398,11 @@ def read_configuration(path):
             result[paths_name] = \
                 [os.path.normpath(os.path.join(cfgdir, p)) for p in paths]
 
+    if 'userdata' in cfg.sections():
+        result['userdata'] = cfg.items('userdata')
+    else:
+        result['userdata'] = []
+
     return result
 
 
@@ -454,6 +465,7 @@ class Configuration(object):
         summary=True,
         junit=False,
         stage=None,
+        userdata=[],
         # -- SPECIAL:
         default_format="pretty",   # -- Used when no formatters are configured.
         scenario_outline_annotation_schema=u"{name} -- @{row.id} {examples.name}"
@@ -561,6 +573,8 @@ class Configuration(object):
             # -- USE ENVIRONMENT-VARIABLE, if stage is undefined.
             self.stage = os.environ.get("BEHAVE_STAGE", None)
         self.setup_stage(self.stage)
+
+        self.userdata = UserData(self.userdata)
         self.setup_model()
 
     def collect_unknown_formats(self):
@@ -662,3 +676,17 @@ class Configuration(object):
             environment_file = prefix + environment_file
         self.steps_dir = steps_dir
         self.environment_file = environment_file
+
+
+class UserData(object):
+    def __init__(self, userdata):
+        for entry in userdata:
+            name, value = self.parse_entry(entry)
+            self.__setattr__(name, value)
+
+    def parse_entry(self, entry):
+        if isinstance(entry, str):
+            parts = entry.split("=", 1)
+            return parts[0], parts[1] if len(parts) > 1 else ""
+
+        return entry