+++ CHANGES.rst
@@ -20,6 +20,7 @@ NEWS and CHANGES:
     * ScenarioOutline: Annotates name with row.id, ... to better represent row.
     * NEW: Active Tags, see docs (`New and Noteworthy`_).
     * NEW: Test stages, see docs (`New and Noteworthy`_).
+    * NEW: Custom userdata config, see docs (`New and Noteworthy`_).
 
   - Model:
 
@@ -34,6 +35,7 @@ ENHANCEMENTS:
   * issue #240: Test Stages with different step implementations (provided by: attilammagyar, jenisys)
   * issue #238: Allow to skip scenario in step function (provided by: hotgloupi, jenisys)
   * issue #228: Exclude scenario fron run (provided by: jdeppe, jenisys)
+  * issue #227: Add a way to add command line options to behave (provided by: mbhinder, attilammagyar, jenisys)
 
 FIXED: